refactor: request/resource, mapper/Mapper, service/serviceImpl

+++ spring-breeze-erp-v3/back/bootSBerp/src/main/java/com/sb/erp/res/dto/request/ResSearchRequest.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 
 @Getter @Setter
 public class ResSearchRequest {
-	private Integer comId;
+	private Long comId;
 	private String keyword;
 	private String resType;
 	private String resStatus;
